Vim settings: skip lost+found and encrypted dirs, make Y default.

+++ setup.sh
@@ -438,9 +438,9 @@ ask_to_proceed "$step"
 
 # VIM SETTINGS ========================================
 
-echo -n "Should we give each user Dan's vim settings? [N|y]: "
+echo -n "Should we give each user Dan's vim settings? [Y|n]: "
 read -e
-if [[ "$REPLY" == y || "$REPLY" == Y ]] ; then
+if [[ -z "$REPLY" || "$REPLY" == y || "$REPLY" == Y ]] ; then
     cd
     git clone git://github.com/convissor/vim-settings.git
     cd vim-settings
@@ -449,6 +449,14 @@ if [[ "$REPLY" == y || "$REPLY" == Y ]] ; then
     dirs=`ls /home`
 
     for dir in $dirs ; do
+        if [ "$dir" == "lost+found" ] ; then
+            continue
+        fi
+        if [ -f "/home/$dir/Access-Your-Private-Data.desktop" ] ; then
+            # Encrypted home directory.
+            continue
+        fi
+
         cd "/home/$dir"
         cp -R /root/vim-settings .
         cd vim-settings
